Moving components to their own files
