+++ src/components/ResetButton.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+export default function ResetButton(props) {
+  return (
+    <button
+      onClick={() => {
+        props.resetGame();
+      }}
+    >
+      Reset Game
+    </button>
+  );
+}
